Move TestTouchInputAfterTouchingComposeArea to separate test scene
